Move code to one file (app.js)+(h.js)→index.js.
- One tiny blob. This helps convey the idea, just one file!
- Save moarr bytes.
- Easier to serve as an ES module.
