--- src/h.js
@@ -1,25 +1,0 @@
-var i
-var stack = []
-
-export function h(type, props) {
-  var node
-  var children = []
-
-  for (i = arguments.length; i-- > 2; ) {
-    stack.push(arguments[i])
-  }
-
-  while (stack.length) {
-    if (Array.isArray((node = stack.pop()))) {
-      for (i = node.length; i--; ) {
-        stack.push(node[i])
-      }
-    } else if (node != null && node !== true && node !== false) {
-      children.push(typeof node === "number" ? (node = node + "") : node)
-    }
-  }
-
-  return typeof type === "string"
-    ? { type: type, props: props || {}, children: children }
-    : type(props || {}, children)
-}
